+++ src/placeholder.rs
@@ -24,10 +24,10 @@ use Cap;
 use Event;
 use Handleable;
 use Id;
-use MetaEvent;
 use Object;
 use Renderable;
 use Renderer;
+use UiEvents;
 use Widget;
 
 
@@ -60,7 +60,7 @@ impl Object for Placeholder {
 }
 
 impl Handleable for Placeholder {
-  fn handle(&mut self, _event: Event, _cap: &mut Cap) -> Option<MetaEvent> {
+  fn handle(&mut self, _event: Event, _cap: &mut Cap) -> Option<UiEvents> {
     unreachable!()
   }
 }